requirements.txt
matplotlib
tqdm
pyyaml
wandb
timm==1.0.11
pillow==11.0.0
tabulate
fvcore
exifread==3.0.0
pyzmq
safetensors
psutil
pyturbojpeg==1.7.7
prettytable
rgbt
k-means-constrained==0.7.6
numpy==2.1.3
opencv-python==4.10.0.84